dags/requirements.txt
-c https://raw.githubusercontent.com/apache/airflow/constraints-1.10.12/constraints-3.8.txt
ethereum-etl-bitski-patched==2.1.2
apache-airflow==1.10.12
google-cloud-storage==1.30.0
httplib2==0.18.1
google_auth_httplib2==0.0.4
google-cloud-bigquery==1.26.1
google-api-python-client==1.10.0
pandas-gbq==0.13.2
pandas==1.1.0
